Move file formats to separate app
Issue #1820
