+++ weblate/trans/management/commands/import_project.py
@@ -29,7 +29,7 @@ from django.core.management.base import BaseCommand, CommandError
 from weblate.lang.models import Language
 from weblate.trans.models import Component, Project
 from weblate.trans.discovery import ComponentDiscovery
-from weblate.trans.formats import FILE_FORMATS
+from weblate.formats import FILE_FORMATS
 from weblate.trans.util import is_repo_link
 from weblate.trans.vcs import VCS_REGISTRY
 from weblate.logger import LOGGER
